Fixing frontend math. With guest star: Frank.

+++ backend/models/procurement_shops.py
@@ -1,7 +1,7 @@
 """Module containing the Procurement Shop model."""
 
 from models.base import BaseModel
-from sqlalchemy import Column, Float, Integer, String
+from sqlalchemy import Column, Float, Integer, String, Numeric
 from sqlalchemy.orm import relationship
 
 
@@ -13,7 +13,7 @@ class ProcurementShop(BaseModel):
     id = Column(Integer, primary_key=True)
     name = Column(String, nullable=False)
     abbr = Column(String, nullable=False)
-    fee = Column(Float, default=0.0)
+    fee = Column(Numeric(12, 5), default=0.0)
     agreements = relationship("Agreement", back_populates="procurement_shop")
 
     @BaseModel.display_name.getter
